Moves common functions into githubAPI.js. Not yet working!
