+++ recommend.js
@@ -1,6 +1,7 @@
 const request = require('request');
 const fs = require('fs');
 const dotenv = require('dotenv').config('.env');
+const githubAPI = require('./githubAPI.js');
 
 const apiCredentials = {
   username: process.env.GITHUB_USER,
